requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==1.4.10
annotated-types==0.7.0
asgiref==3.9.2
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
easydict==1.13
facenet-pytorch==2.6.0
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h5py==3.14.0
holidays==0.81
humanfriendly==10.0
idna==3.10
imageio==2.37.0
inflection==0.5.1
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.5
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==10.2.0
prettytable==3.16.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.36.2
requests==2.32.5
retina-face==0.0.17
rich==14.1.0
rpds-py==0.27.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stringzilla==4.1.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.9.30
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.11.0
wrapt==1.17.3
